Fix I_MINT sign-extension and constmap constant truncation on 64-bit
I_MINT (MOSTNEG INT) used (unsigned int)0x80000000 which zero-extends
to 0x0000000080000000 (positive) instead of (int)0x80000000 which
sign-extends to 0xFFFFFFFF80000000 (negative). This caused INT64
comparisons with MOSTNEG INT to treat -2147483648 as positive.

Also fix constmap_new to preserve full intptr_t constant values
instead of truncating via (int) cast.

Fixes cgtest22 B1333a6. Reduces cgtest22 to 3 remaining errors.

+++ tools/tranx86/tstack.c
@@ -490,7 +490,7 @@ void constmap_new (int reg, int consttype, intptr_t constval, ins_chain *ins)
 		tmp->c_val = labs[0];
 		tmp->c_val2 = labs[1];
 	} else {
-		tmp->c_val = (int)constval;
+		tmp->c_val = constval;  /* preserve full intptr_t value for 64-bit constants */
 		tmp->c_val2 = 0;
 	}
 	tmp->load_ins = ins;
